feat(#106): remove the puzzle for #106 issue
